refactor(react): restructure code
