+++ apps/docs/app/sample/create/page.tsx
@@ -1,0 +1,9 @@
+import { Create } from "@honohub/react";
+
+export default function CreatePage() {
+  return (
+    <div>
+      <Create />
+    </div>
+  );
+}
