+++ apps/docs/app/sample/list/page.tsx
@@ -1,0 +1,5 @@
+import { ListPage } from "@honohub/react";
+
+export default function ListSamplePage() {
+  return <ListPage />;
+}
