+++ apps/docs/app/sample/page.tsx
@@ -1,29 +1,5 @@
-import Link from "next/link";
-import React from "react";
-
-const DATA = [
-  {
-    label: "List",
-    href: "/sample/list",
-  },
-  {
-    label: "Create",
-    href: "/sample/create",
-  },
-];
+import { AdminPage } from "@honohub/react";
 
 export default function SamplePage() {
-  return (
-    <div className="h-full space-y-5">
-      <h1 className="text-3xl font-bold">Collections</h1>
-      <div className="grid grid-cols-3 gap-4">
-        {DATA.map(({ label, href }, index) => (
-          // biome-ignore lint/suspicious/noArrayIndexKey: <explanation>
-          <Link key={index} href={href} className="p-4 border rounded-md">
-            <p>{label}</p>
-          </Link>
-        ))}
-      </div>
-    </div>
-  );
+  return <AdminPage />;
 }
