+++ packages/react/src/Pages/Create.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function Create() {
+  return <div>Create</div>;
+}
