+++ packages/react/src/Pages/List/cells/ClipboardCell.tsx
