+++ packages/react/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./lib/react";
-export * from "./Components";
+export * from "./Pages";
+export { Breadcrumbs, Header } from "./Components";
