--- packages/react/src/lib/react.tsx
@@ -1,9 +1,0 @@
-export function React() {
-  return (
-    <div>
-      <h1>Welcome to React!</h1>
-    </div>
-  );
-}
-
-export default React;
